+++ ingredient_parser/_common.py
@@ -14,8 +14,8 @@ import nltk
 
 SUPPORTED_LANGUAGES = ["en"]
 
-# Regex pattern for matching a numeric range e.g. 1-2, 2-3.
-RANGE_PATTERN = re.compile(r"^\d+\s*[\-]\d+$")
+# Regex pattern for matching a numeric range e.g. 1-2, 2-3, #1$2-1#3$4.
+RANGE_PATTERN = re.compile(r"^[\d\#\$]+\s*[\-][\d\#\$]+$")
 
 
 def consume(iterator: Iterator, n: int | None) -> None: