+++ ingredient_parser/en/postprocess.py
@@ -88,7 +88,7 @@ class PostProcessor:
     string_units : bool
         If True, return all IngredientAmount units as strings.
         If False, convert IngredientAmount units to pint.Unit objects where possible.
-        Dfault is False.
+        Default is False.
     imperial_units : bool
         If True, use imperial units instead of US customary units for pint.Unit objects
         for the the following units: fluid ounce, cup, pint, quart, gallon.
@@ -108,6 +108,7 @@ class PostProcessor:
         discard_isolated_stop_words: bool = True,
         string_units: bool = False,
         imperial_units: bool = False,
+        quantity_fractions: bool = False,
     ):
         self.sentence = sentence
         self.tokens = tokens
@@ -116,6 +117,7 @@ class PostProcessor:
         self.discard_isolated_stop_words = discard_isolated_stop_words
         self.string_units = string_units
         self.imperial_units = imperial_units
+        self.quantity_fractions = quantity_fractions
         self.consumed = []
 
     def __repr__(self) -> str:
@@ -184,20 +186,21 @@ class PostProcessor:
         """
         # Select indices of tokens, labels and scores for selected_label
         # Do not include tokens, labels and scores in self.consumed
-        idx = [
+        label_idx = [
             i
             for i, label in enumerate(self.labels)
             if label in [selected_label, "PUNC"] and i not in self.consumed
         ]
 
         # If idx is empty or all the selected idx are PUNC, return None
-        if not idx or all(self.labels[i] == "PUNC" for i in idx):
+        if not label_idx or all(self.labels[i] == "PUNC" for i in label_idx):
             return None
 
         # Join consecutive tokens together and average their score
         parts = []
         confidence_parts = []
-        for group in group_consecutive_idx(idx):
+        starting_index = label_idx[-1]
+        for group in group_consecutive_idx(label_idx):
             idx = list(group)
             idx = self._remove_invalid_indices(idx)
 
@@ -208,13 +211,14 @@ class PostProcessor:
             joined = " ".join([self.tokens[i] for i in idx])
             confidence = mean([self.scores[i] for i in idx])
 
-            if self.discard_isolated_stop_words and joined in STOP_WORDS:
+            if self.discard_isolated_stop_words and joined.lower() in STOP_WORDS:
                 # Skip part if it's a stop word
                 continue
 
             self.consumed.extend(idx)
             parts.append(joined)
             confidence_parts.append(confidence)
+            starting_index = min(starting_index, idx[0])
 
         # Find the indices of the joined tokens list where the element
         # is the same as the previous element in the list.
@@ -238,6 +242,7 @@ class PostProcessor:
         return IngredientText(
             text=text,
             confidence=round(mean(confidence_parts), 6),
+            starting_index=starting_index,
         )
 
     def _postprocess_amounts(self) -> list[IngredientAmount]:
@@ -478,7 +483,7 @@ class PostProcessor:
 
         QTY_idx = [i for i, label in enumerate(self.labels) if label == "QTY"]
 
-        # Find any cases where a group of consecutuve QTY tokens can be collapsed into
+        # Find any cases where a group of consecutive QTY tokens can be collapsed into
         # a single token. Modify the first token and score in the group and mark all
         # others in group for deletion.
         idx_to_remove = []
@@ -546,7 +551,7 @@ class PostProcessor:
         Parameters
         ----------
         idx : list[int]
-            List of indices of the tokens/labels/scores in the full tokenizsed sentence
+            List of indices of the tokens/labels/scores in the full tokenized sentence
         tokens : list[str]
             Tokens for input sentence
         labels : list[str]
@@ -572,6 +577,7 @@ class PostProcessor:
             "bag",
             "block",
             "box",
+            "bucket",
             "can",
             "container",
             "envelope",
@@ -616,6 +622,7 @@ class PostProcessor:
                         APPROXIMATE=self._is_approximate(match[0], tokens, labels, idx),
                         string_units=self.string_units,
                         imperial_units=self.imperial_units,
+                        quantity_fractions=self.quantity_fractions,
                     )
                     amounts.append(first)
                     # Pop the first and last items from the list of matching indices
@@ -642,6 +649,7 @@ class PostProcessor:
                             APPROXIMATE=first.APPROXIMATE,
                             string_units=self.string_units,
                             imperial_units=self.imperial_units,
+                            quantity_fractions=self.quantity_fractions,
                         )
                         amounts.append(amount)
 
@@ -659,7 +667,7 @@ class PostProcessor:
         * 1 pint 2 fl oz
         * 2 cups plus 1 tablespoon
 
-        Return a compositive amount object made from the adjacent amounts.
+        Return a composite amount object made from the adjacent amounts.
 
         For example, for the sentence: 1 lb 2 oz ...; the composite amount is:
         CompositeAmount(
@@ -673,7 +681,7 @@ class PostProcessor:
         Parameters
         ----------
         idx : list[int]
-            List of indices of the tokens/labels/scores in the full tokenizsed sentence
+            List of indices of the tokens/labels/scores in the full tokenized sentence
         tokens : list[str]
             Tokens for input sentence
         labels : list[str]
@@ -693,6 +701,7 @@ class PostProcessor:
         patterns = {
             "ptfloz": {
                 "pattern": ["QTY", "UNIT", "QTY", "UNIT", "UNIT"],
+                "conjunction": None,
                 "start1": 0,
                 "start2": 2,
                 "join": "",
@@ -700,6 +709,7 @@ class PostProcessor:
             },
             "lboz": {
                 "pattern": ["QTY", "UNIT", "QTY", "UNIT"],
+                "conjunction": None,
                 "start1": 0,
                 "start2": 2,
                 "join": "",
@@ -707,13 +717,31 @@ class PostProcessor:
             },
             "plus": {
                 "pattern": ["QTY", "UNIT", "COMMENT", "QTY", "UNIT"],
+                "conjunction": "plus",
                 "start1": 0,
                 "start2": 3,
                 "join": " plus ",
                 "subtractive": False,
             },
+            "and": {
+                "pattern": ["QTY", "UNIT", "COMMENT", "QTY", "UNIT"],
+                "conjunction": "and",
+                "start1": 0,
+                "start2": 3,
+                "join": " and ",
+                "subtractive": False,
+            },
+            "plus_punc": {
+                "pattern": ["QTY", "UNIT", "PUNC", "QTY", "UNIT"],
+                "conjunction": "+",
+                "start1": 0,
+                "start2": 3,
+                "join": " + ",
+                "subtractive": False,
+            },
             "minus": {
                 "pattern": ["QTY", "UNIT", "COMMENT", "QTY", "UNIT"],
+                "conjunction": "minus",
                 "start1": 0,
                 "start2": 3,
                 "join": " minus ",
@@ -721,6 +749,7 @@ class PostProcessor:
             },
             "less": {
                 "pattern": ["QTY", "UNIT", "COMMENT", "QTY", "UNIT"],
+                "conjunction": "less",
                 "start1": 0,
                 "start2": 3,
                 "join": " minus ",
@@ -745,7 +774,7 @@ class PostProcessor:
                 labels, pattern, ignore_other_labels=False
             ):
                 # Check if match fits with "ptfloz" or "lboz" pattern constraints
-                if pattern_name in ["pltfloz", "lboz"]:
+                if pattern_name in ["ptfloz", "lboz"]:
                     first_unit = tokens[match[start1 + 1]]
                     last_unit = tokens[match[-1]]
                     if (
@@ -756,12 +785,10 @@ class PostProcessor:
                         # ptfloz or lboz patterns, so skip
                         continue
 
-                # Check if match fits with plus/minus/less pattern constraints
-                if pattern_name in ["plus", "minus", "less"]:
-                    if tokens[match[2]].lower() != pattern_name:
-                        # Middle token with COMMENT label is not same as name of pattern
-                        # so skip.
-                        continue
+                # For other patterns, check if third token in match matches conjunction
+                # and skip if not.
+                elif tokens[match[2]].lower() != pattern_info["conjunction"]:
+                    continue
 
                 # First amount
                 quantity_1 = tokens[match[start1]]
@@ -777,6 +804,7 @@ class PostProcessor:
                     starting_index=idx[match[start1]],
                     string_units=self.string_units,
                     imperial_units=self.imperial_units,
+                    quantity_fractions=self.quantity_fractions,
                 )
 
                 # Second amount
@@ -793,6 +821,7 @@ class PostProcessor:
                     starting_index=idx[match[start2]],
                     string_units=self.string_units,
                     imperial_units=self.imperial_units,
+                    quantity_fractions=self.quantity_fractions,
                 )
 
                 composite_amounts.append(
@@ -891,7 +920,7 @@ class PostProcessor:
         Parameters
         ----------
         idx : list[int]
-            List of indices of the tokens/labels/scores in the full tokenizsed sentence
+            List of indices of the tokens/labels/scores in the full tokenized sentence
         tokens : list[str]
             Tokens for input sentence
         labels : list[str]
@@ -985,6 +1014,7 @@ class PostProcessor:
                     SINGULAR=amount.SINGULAR,
                     string_units=self.string_units,
                     imperial_units=self.imperial_units,
+                    quantity_fractions=self.quantity_fractions,
                 )
             )
 
@@ -1009,7 +1039,7 @@ class PostProcessor:
         labels : list[str]
             List of all token labels
         idx : list[int]
-            List of indices of the tokens/labels/scores in the full tokenizsed sentence
+            List of indices of the tokens/labels/scores in the full tokenized sentence
 
         Returns
         -------
@@ -1075,7 +1105,7 @@ class PostProcessor:
         labels : list[str]
             List of all token labels
         idx : list[int]
-            List of indices of the tokens/labels/scores in the full tokenizsed sentence
+            List of indices of the tokens/labels/scores in the full tokenized sentence
 
         Returns
         -------
@@ -1104,7 +1134,7 @@ class PostProcessor:
         if i == len(tokens) - 2:
             return False
 
-        # Case where the amonut is in brackets
+        # Case where the amount is in brackets
         if (
             labels[i] == "UNIT"
             and tokens[i + 1] in [")", "]"]
@@ -1138,7 +1168,7 @@ class PostProcessor:
         labels : list[str]
             List of all token labels
         idx : list[int]
-            List of indices of the tokens/labels/scores in the full tokenizsed sentence
+            List of indices of the tokens/labels/scores in the full tokenized sentence
 
         Returns
         -------